serverless-deploy/mlfunction/requirements.txt
azure-functions
tensorflow==1.14
Pillow
requests